Locate test helpers into separate folder
