+++ api/test/create_content.js
@@ -2,8 +2,8 @@
 
 // tests for create_content
 // Generated by serverless-mocha-plugin
-require('./test_provider_configurator').configure();
-const DatabaseCleaner = require('./database_cleaner');
+require('./helpers/TestProviderConfigurator').configure();
+const DatabaseCleaner = require('./helpers/DatabaseCleaner');
 
 const mod = require('../content/create.js');
 const mochaPlugin = require('serverless-mocha-plugin');
