+++ api/test/create_venue.js
@@ -2,8 +2,8 @@
 
 // tests for create_venue
 // Generated by serverless-mocha-plugin
-require('./test_provider_configurator').configure();
-const DatabaseCleaner = require('./database_cleaner');
+require('./helpers/TestProviderConfigurator').configure();
+const DatabaseCleaner = require('./helpers/DatabaseCleaner');
 
 const lambda = require('../venues/create.js');
 const mochaPlugin = require('serverless-mocha-plugin');
