+++ api/test/helpers/TestProviderConfigurator.js
