+++ api/test/update_venue.js
@@ -2,8 +2,8 @@
 
 // tests for update_venue
 // Generated by serverless-mocha-plugin
-require('./test_provider_configurator').configure();
-const DatabaseCleaner = require('./database_cleaner');
+require('./helpers/TestProviderConfigurator').configure();
+const DatabaseCleaner = require('./helpers/DatabaseCleaner');
 
 const mod = require('../venues/update.js');
 const createFunction = require('../venues/create.js');
