+++ api/test/venues_list.js
@@ -2,8 +2,8 @@
 
 // tests for venues_list
 // Generated by serverless-mocha-plugin
-require('./test_provider_configurator').configure();
-const DatabaseCleaner = require('./database_cleaner');
+require('./helpers/TestProviderConfigurator').configure();
+const DatabaseCleaner = require('./helpers/DatabaseCleaner');
 
 const getAllFunction = require('../venues/list.js');
 const createFunction = require('../venues/create.js');
